code structure separation api
